feat(key-manager): implement JWE functionality directly in `key-manager`
The KMS is only used to compute shared secret, it doesn't need to deal with JWE directly
fixes #556
